u-boot-fslc: deploy dtb files for mx8m machines
U-Boot DTB files need to be additionally deployed in order to be picked
up later by imx-boot component to produce boot container.

As the boot container is required to be provided as a boot binary file
for mx8m derivatives, deploy DTB files to a known location within
deployment folder in addition to u-boot binary files.

Signed-off-by: Andrey Zhizhikin <andrey.z@gmail.com>