imx-atf: allow to be used with mainline bsp for mx8m machines
Since ATF is required for i.MX8M series to boot the SOC, allow it to be
built when use-mainline-bsp is set.

Move ATF build platform definition to target machine configuration
files, since MACHINEOVERRIDES extender filters out the override that is
used to identify it in the recipe. Therefore, for those mx8m machines
that are ready to be built with mainline BSP - ATF build platform is
explicitly defined in the machine config in order to be properly set for
ATF build.

Signed-off-by: Andrey Zhizhikin <andrey.z@gmail.com>