imx-atf: change compatibility to boot-container override
Mainline BSP in general is not compatible with NXP version of ATF
provided. It is required only for mx8m series, where a boot container
is built as a replacements of imx-boot target.

Change compatibility from use-mainline-bsp to boot-container to limit
build only to those machines, which require the boot container.

Signed-off-by: Andrey Zhizhikin <andrey.z@gmail.com>